Global Exception Handling added

+++ ticket-system-web/src/main/java/com/tsystems/javaschool/vm/dao/CommonDAO.java
@@ -1,9 +1,9 @@
 package com.tsystems.javaschool.vm.dao;
 
 import com.tsystems.javaschool.vm.domain.SBBEntity;
+import com.tsystems.javaschool.vm.exception.EntityNotFoundException;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityTransaction;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 import java.util.List;
@@ -14,32 +14,27 @@ public abstract class CommonDAO<E extends SBBEntity>{
 
     @PersistenceContext(unitName = "SBBPU")
     protected EntityManager entityManager;
-    
-    
 
     protected CommonDAO(Class<E> entityClass) {
         this.entityClass = entityClass;
-    }
-
-    public EntityTransaction getTransaction() {
-        return entityManager.getTransaction();
-    }
-
-    public Query createQuery(String queryString) {
-        return entityManager.createQuery(queryString);
     }
 
     public void create(E entity) {
         entityManager.persist(entity);
     }
 
-    public E findById(Long id) {
-        return entityManager.find(entityClass, id);
+    public E findById(Long id) throws EntityNotFoundException {
+        E entity = entityManager.find(entityClass, id);
+        if (entity == null) {
+            throw new EntityNotFoundException(entityClass.getCanonicalName() + " not found; id: " + id);
+        }
+        return entity;
     }
 
     public void update(E entity) {
-        entityManager.merge(entity);
-
+        if (entity.getId() > 0) {
+            entityManager.merge(entity);
+        }
     }
 
     public void detach(E entity) {
@@ -51,9 +46,11 @@ public abstract class CommonDAO<E extends SBBEntity>{
     }
 
     public void delete(Long id) {
-        E entity = entityManager.find(entityClass, id);
-        if (entity != null) {
-            entityManager.remove(entity);
+        if (id > 0) {
+            E entity = entityManager.find(entityClass, id);
+            if (entity != null) {
+                entityManager.remove(entity);
+            }
         }
     }
